backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
numpy
scikit-learn
faiss-cpu
sentence-transformers
langchain
langchain-community
python-multipart
Pillow
torch
torchvision
tqdm
jinja2
